Update List Paket dan Manajemen Login,Register dan Logout

+++ app/Http/Controllers/API/Pilihan_Paket/InputPesananController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Api;
+namespace App\Http\Controllers\API\Pilihan_Paket;
 
 use App\Http\Controllers\Controller;
 use App\Models\InputPesanan;
@@ -24,8 +24,9 @@ class InputPesananController extends Controller
             'Berat_Barang' => 'required',
             'Alamat_Tujuan' => 'required',
             'status_pembayaran' => 'required',
-            'Nama_Kurir' => 'required',
-            'nama' => 'required',
+            // 'Nama_Kurir' => 'required',
+            // 'nama' => 'required',
+            // 'Nama_Paket' => 'required',
         ]);
 
         $inputPesanan = InputPesanan::create($validatedData);
@@ -42,25 +43,32 @@ class InputPesananController extends Controller
     public function update(Request $request, $id)
     {
         $request->validate([
-            'Nama_Barang' => 'required',
-            'Generate_Resi' => 'required',
-            'Berat_Barang' => 'required',
-            'Alamat_Tujuan' => 'required',
-            'status_pembayaran' => 'required',
-            'Nama_Kurir' => 'required',
-            'nama' => 'required',
+            'Nama_Barang',
+            'Generate_Resi',
+            'Berat_Barang' ,
+            'Alamat_Tujuan',
+            'status_pembayaran',
+            // 'Nama_Kurir' => 'required',
+            // 'nama' => 'required',
+            // 'Nama_Paket' => 'required',
         ]);
 
-        $id->update($request->all());
+        $inputPesanan = InputPesanan::findOrFail($id);
+        $inputPesanan->update($request->all());
         
-        return response()->json(['message' => 'Data Berhasil Di Ubah', 'Data' => $id], 200);
+        return response()->json(['message' => 'Data Berhasil Di Ubah', 'Data' => $id , $inputPesanan], 200);
     }
 
     public function destroy($id)
     {
-        $inputPesanan = InputPesanan::findOrFail($id);
+        $inputPesanan = InputPesanan::find($id);
+    
+        if (!$inputPesanan) {
+            return response()->json(['message' => 'Data Tidak Ditemukan!'], 404);
+        }
+    
         $inputPesanan->delete();
-        
+    
         return response()->json(['message' => 'Data Berhasil Dihapus'], 200);
     }
 }
